Updated to v0.4
Added multiple filed extractions for the email header and file
attachments
Added support for decoding unicode characters in other languages or and
removing the unicode identifier in the header.
Added support for indexing some file types even if the content-type is
not set correctly. (as with Microsoft sending some files as binaries
instead of text)
Added fundamental code to support indexing of attachment as a
configurable option by the user.